backend/pyproject.toml
[build-system]
requires = ["setuptools ~= 80.3.1"]
build-backend = "setuptools.build_meta"

[project]
name = "crypto"
version = "0.1.0"
description = "Crypto Guesser"
readme = "README.md"
requires-python = ">=3.9"

dependencies = [
    "base58",
    "django",
    "djangorestframework",
    "drf-nested-routers",
    "ecdsa",
    "sympy",
]

[project.optional-dependencies]
dev = ["pytest", "pytest-django", "black", "isort", "flake8", "flake8-bugbear", "Flake8-pyproject", "pre-commit"]

[tool.setuptools]
zip-safe = false

[tool.setuptools.packages.find]
where = ["project"]

[tool.isort]
profile = "black"

[tool.flake8]
max-line-length = 80
extend-select = "B950"
extend-ignore = ["E203","E501","E701"]

[tool.pytest.ini_options]
minversion = "6.0"
addopts = "--strict-markers"
DJANGO_SETTINGS_MODULE = "project.settings"
python_files = ["tests.py", "test_*.py", "*_tests.py"]
